requirements.txt
# Core dependencies
PySimpleGUI>=5.0.0
onnxruntime>=1.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
